Add createDomainWithActions tests
